% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepstate_compile_AFL.R
\name{deepstate_compile_tools}
\alias{deepstate_compile_tools}
\title{creates afl fuzzer for given functions in package}
\usage{
deepstate_compile_tools(path)
}
\arguments{
\item{path}{to the package to test}
}
\description{
creates afl fuzzer for given functions in package
}
