% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepstate_compile_tools.R
\name{deepstate_make_afl}
\alias{deepstate_make_afl}
\title{creates afl fuzzer for given functions in package}
\usage{
deepstate_make_afl()
}
\description{
creates afl fuzzer for given functions in package
}
