% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepstate_Angora.R
\name{deepstate_make_angora}
\alias{deepstate_make_angora}
\title{creates Angora fuzzer for given functions in package}
\usage{
deepstate_make_angora()
}
\description{
creates Angora fuzzer for given functions in package
}
