% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepstate_Eclipser.R
\name{deepstate_make_eclipser}
\alias{deepstate_make_eclipser}
\title{creates Eclipser fuzzer for given functions in package}
\usage{
deepstate_make_eclipser()
}
\description{
creates Eclipser fuzzer for given functions in package
}
