% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepstate_HonggFuzz.R
\name{deepstate_make_hongg}
\alias{deepstate_make_hongg}
\title{creates Hongg fuzzer for given functions in package}
\usage{
deepstate_make_hongg()
}
\description{
creates Hongg fuzzer for given functions in package
}
