% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepstate_LibFuzzer.R
\name{deepstate_make_libFuzzer}
\alias{deepstate_make_libFuzzer}
\title{creates libfuzzer for given functions in package}
\usage{
deepstate_make_libFuzzer()
}
\description{
creates libfuzzer for given functions in package
}
