% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepstate_AFL.R
\name{deepstate_pkg_create_AFL}
\alias{deepstate_pkg_create_AFL}
\title{creates afl fuzzer specific make for given functions in package}
\usage{
deepstate_pkg_create_AFL(path)
}
\arguments{
\item{path}{to the package to test}
}
\description{
creates afl fuzzer specific make for given functions in package
}
