% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepstate_Eclipser.R
\name{deepstate_pkg_create_Eclipser}
\alias{deepstate_pkg_create_Eclipser}
\title{creates Eclipser specific make files for given functions in package}
\usage{
deepstate_pkg_create_Eclipser(path)
}
\arguments{
\item{path}{to the package to test}
}
\description{
creates Eclipser specific make files for given functions in package
}
