% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepstate_HonggFuzz.R
\name{deepstate_pkg_create_HonggFuzz}
\alias{deepstate_pkg_create_HonggFuzz}
\title{creates hongg fuzzer specific make for given functions in package}
\usage{
deepstate_pkg_create_HonggFuzz(path)
}
\arguments{
\item{path}{to the package to test}
}
\description{
creates hongg fuzzer specific make for given functions in package
}
