% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepstate_LibFuzzer.R
\name{deepstate_pkg_create_LibFuzzer}
\alias{deepstate_pkg_create_LibFuzzer}
\title{creates LibFuzzer specific make files for given functions in package}
\usage{
deepstate_pkg_create_LibFuzzer(path)
}
\arguments{
\item{path}{to the package to test}
}
\description{
creates LibFuzzer specific make files for given functions in package
}
